backend/requirements.txt
alabaster==0.7.16
asgiref==3.8.1
astroid==3.2.4
attrs==23.2.0
Babel==2.15.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-netfields==1.3.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.18.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
idna==3.7
imagesize==1.4.1
inflection==0.5.1
isort==5.13.2
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
netaddr==1.3.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg==3.1.19
psycopg-binary==3.1.19
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.5
PyYAML==6.0.1
referencing==0.35.1
renku-sphinx-theme==0.4.0
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.13.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
